Refactored UniquePersonList to UniqueTaskList
Refactored UniquePersonList to UniqueTaskList under
seedu.address.model.task package
