+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -4,7 +4,7 @@ package seedu.address.model;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 import seedu.address.model.task.ReadOnlyTask;
-import seedu.address.model.task.UniquePersonList;
+import seedu.address.model.task.UniqueTaskList;
 
 import java.util.List;
 
@@ -15,7 +15,7 @@ public interface ReadOnlyAddressBook {
 
     UniqueTagList getUniqueTagList();
 
-    UniquePersonList getUniquePersonList();
+    UniqueTaskList getUniquePersonList();
 
     /**
      * Returns an unmodifiable view of persons list
